// frontend/src/client/types.gen.ts
export type BugCreate = {
    title: string;
    description?: (string | null);
    severity?: BugSeverity;
    status?: BugStatus;
    assignee_id?: (string | null);
    project_id: string;
};

export type BugPublic = {
    id: string;
    title: string;
    description?: (string | null);
    severity: BugSeverity;
    status: BugStatus;
    assignee_id?: (string | null);
    project_id: string;
    created_at?: (string | null);
    updated_at?: (string | null);
};

export type BugSeverity = 'low' | 'medium' | 'high' | 'critical';

export type BugsPublic = {
    data: Array<BugPublic>;
    count: number;
};

export type BugStatus = 'open' | 'in_progress' | 'fixed' | 'closed';